trxcon/l1sched: refactor l1sched_mframe_layout()

+++ src/host/trxcon/include/osmocom/bb/l1sched/l1sched.h
@@ -327,8 +327,8 @@ struct l1sched_state {
 };
 
 extern const struct l1sched_lchan_desc l1sched_lchan_desc[_L1SCHED_CHAN_MAX];
-const struct l1sched_tdma_multiframe *l1sched_mframe_layout(
-	enum gsm_phys_chan_config config, int tn);
+const struct l1sched_tdma_multiframe *
+l1sched_mframe_layout(enum gsm_phys_chan_config config, uint8_t tn);
 
 /* Scheduler management functions */
 struct l1sched_state *l1sched_alloc(void *ctx, const struct l1sched_cfg *cfg, void *priv);
